+++ pysui/sui/sui_clients/transaction_builder.py
@@ -20,6 +20,8 @@ from math import ceil
 from typing import Optional, Set, Union
 from functools import singledispatchmethod
 
+from deprecated.sphinx import versionchanged
+
 from pysui.sui.sui_types import bcs
 from pysui.sui.sui_types.address import SuiAddress
 from pysui.sui.sui_types.scalars import ObjectID, SuiInteger, SuiString
@@ -152,18 +154,7 @@ class ProgrammableTransactionBuilder:
         self.objects_registry.add(key.value.to_address_str())
         return bcs.Argument("Input", out_index)
 
-    # def obj(self, obj_arg: bcs.ObjectArg) -> bcs.Argument:
-    #     """."""
-    #     for index, (key, value) in enumerate(self.inputs.values()):
-    #         # Check callarg (value) content with inbound
-    #         carg_val = value.value
-    #         # If collected is ObjectArg and we have equal values
-    #         if value.index == 1 and carg_val == obj_arg:
-    #             # If both are SharedObject
-    #             if carg_val.index == 1 and obj_arg.index == 1:
-    #                 pass
-
-    def command(self, command_obj: bcs.Command) -> bcs.Argument:
+    def command(self, command_obj: bcs.Command, nresults: int = 1) -> Union[bcs.Argument, list[bcs.Argument]]:
         """command adds a new command to the list of commands.
 
         :param command_obj: The Command type
@@ -173,6 +164,11 @@ class ProgrammableTransactionBuilder:
         """
         out_index = len(self.commands)
         self.commands.append(command_obj)
+        if nresults > 1:
+            nreslist: list[bcs.Argument] = []
+            for nrindex in range(nresults):
+                nreslist.append(bcs.Argument("NestedResult", (out_index, nrindex)))
+            return nreslist
         return bcs.Argument("Result", out_index)
 
     def make_move_vector(
@@ -221,23 +217,19 @@ class ProgrammableTransactionBuilder:
             )
         )
 
+    @versionchanged(version="0.17.0", reason="Extend to take list of amounts")
     def split_coin(
         self,
         from_coin: Union[bcs.Argument, tuple[bcs.BuilderArg, bcs.ObjectArg]],
-        amount: bcs.BuilderArg,
+        amounts: list[bcs.BuilderArg],
     ) -> bcs.Argument:
         """Setup a SplitCoin command and return it's result Argument."""
-        amount_arg = self.input_pure(amount)
+        amounts_arg = [self.input_pure(x) for x in amounts]
         if isinstance(from_coin, bcs.Argument):
             coin_arg = from_coin
         else:
             coin_arg = self.input_obj(*from_coin)
-        return self.command(
-            bcs.Command(
-                "SplitCoin",
-                bcs.SplitCoin(coin_arg, [amount_arg]),
-            )
-        )
+        return self.command(bcs.Command("SplitCoin", bcs.SplitCoin(coin_arg, amounts_arg)), len(amounts_arg))
 
     def merge_coins(
         self,
@@ -278,7 +270,7 @@ class ProgrammableTransactionBuilder:
         """
         reciever_arg = self.input_pure(recipient)
         if amount:
-            coin_arg = self.split_coin(from_coin=from_coin, amount=amount)
+            coin_arg = self.split_coin(from_coin=from_coin, amounts=amount)
         else:
             coin_arg = self.input_obj(*from_coin)
         return self.command(bcs.Command("TransferObjects", bcs.TransferObjects([coin_arg], reciever_arg)))