+++ pysui/sui/sui_utils.py
@@ -18,6 +18,7 @@ import os
 import base64
 import binascii
 import subprocess
+import hashlib
 from dataclasses import dataclass
 from pathlib import Path
 from types import NoneType
@@ -25,6 +26,7 @@ from typing import Any, Union
 import base58
 import yaml
 from dataclasses_json import DataClassJsonMixin
+from deprecated.sphinx import versionchanged, versionadded
 from pysui.sui.sui_constants import (
     DEFAULT_DEVNET_PATH_STRING,
     PYSUI_EXEC_ENV,
@@ -65,14 +67,16 @@ _UNPUBLISHED: str = "00000000000000000000000000000000000000000000000000000000000
 
 
 @dataclass
+@versionchanged(version="0.17.0", reason="Added the package digest that matches chain digest.")
 class CompiledPackage:
     """Ease of compilation information dataclass."""
 
     project_name: str
     project_id: str
-    project_digest: bytes
+    project_source_digest: bytes
     dependencies: list[str]
     compiled_modules: list[SuiString] = None
+    package_digest: bytes = None
 
 
 def _compile_project(path_to_package: Path, skip_git_dependencies: bool) -> Union[Path, SuiException]:
@@ -133,6 +137,28 @@ def _build_dep_info(build_path: str) -> Union[CompiledPackage, Exception]:
     raise ValueError("Corrupt publish build information")
 
 
+@versionadded(version="0.17.0", reason="Added true package hash (digest) to CompiledPacakge.")
+def _package_digest(package: CompiledPackage, readers: list[ModuleReader]) -> None:
+    """Converts compiled module bytes for publishing and digest calculation."""
+    mod_strs: list = []
+    all_bytes: list = []
+    # Get the bytes for digest and string for publishing
+    for mod_bytes in readers:
+        mr_bytes = mod_bytes.reader.getvalue()
+        all_bytes.append(mr_bytes)
+        mod_strs.append(SuiString(base64.b64encode(mr_bytes).decode()))
+    for dep_str in package.dependencies:
+        all_bytes.append(binascii.unhexlify(dep_str[2:]))
+
+    all_bytes.sort()
+    hasher = hashlib.blake2b(digest_size=32)
+    for bblock in all_bytes:
+        hasher.update(bblock)
+    package.package_digest = hasher.digest()
+    package.compiled_modules = mod_strs
+
+
+@versionchanged(version="0.17.0", reason="Added the package digest that matches chain digest.")
 def publish_build(
     path_to_package: Path, include_unpublished: bool = False, skip_git_dependencie: bool = False
 ) -> Union[CompiledPackage, Exception]:
@@ -154,9 +180,8 @@ def publish_build(
 
     # Construct initial package
     cpackage = _build_dep_info(build_subdir[0].path)
-    # Get module bytes as base64 strings
-    module_readers: list[ModuleReader] = _modules_bytes(byte_modules)
-    cpackage.compiled_modules = [SuiString(base64.b64encode(mr.reader.getvalue()).decode()) for mr in module_readers]
+    # Set module bytes as base64 strings and generate package digest
+    _package_digest(cpackage, _modules_bytes(byte_modules))
     return cpackage
 
 
@@ -336,6 +361,8 @@ def as_sui_string(in_data: Any) -> Union[SuiString, ValueError]:
         result = SuiString(str(in_data))
     elif issubclass(type(in_data), SuiAddress):
         result = SuiString(in_data.identifier.value)
+    elif isinstance(in_data, SuiNullType):
+        result = in_data
     if not result:
         raise ValueError(f"Can not get SuiString from {in_data} with type {type(in_data)}")
     return result